+++ continuous/state/MorphologicalComputation.go
@@ -1,17 +1,17 @@
 package state
 
 // MorphologicalComputationW [...]
-func MorphologicalComputationW(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationW(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int, k int, eta bool) []float64 {
 	return FrenzelPompe(w2w1a1, w2Indices, w1Indices, a1Indices, k, eta)
 }
 
 // MorphologicalComputationA [...]
-func MorphologicalComputationA(w2a1w1 [][]float64, w2Indices, a1Indices, w1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationA(w2a1w1 [][]float64, w2Indices, a1Indices, w1Indices []int, k int, eta bool) []float64 {
 	return FrenzelPompe(w2a1w1, w2Indices, a1Indices, w1Indices, k, eta)
 }
 
 // MorphologicalComputationCW1 [...]
-func MorphologicalComputationCW1(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationCW1(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int, k int, eta bool) []float64 {
 	r1 := KraskovStoegbauerGrassberger1(w2w1a1, w2Indices, w1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger1(w2w1a1, w2Indices, a1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -22,7 +22,7 @@ func MorphologicalComputationCW1(w2w1a1 [][]float64, w2Indices, w1Indices, a1Ind
 }
 
 // MorphologicalComputationCW2 [...]
-func MorphologicalComputationCW2(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationCW2(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int, k int, eta bool) []float64 {
 	r1 := KraskovStoegbauerGrassberger2(w2w1a1, w2Indices, w1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger2(w2w1a1, w2Indices, a1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -33,7 +33,7 @@ func MorphologicalComputationCW2(w2w1a1 [][]float64, w2Indices, w1Indices, a1Ind
 }
 
 // MorphologicalComputationWA1 = I(W;{W,A}) - I(W';A)
-func MorphologicalComputationWA1(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationWA1(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int, k int, eta bool) []float64 {
 	r1 := FrenzelPompe(w2w1a1, w2Indices, w1Indices, a1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger1(w2w1a1, w2Indices, a1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -44,7 +44,7 @@ func MorphologicalComputationWA1(w2w1a1 [][]float64, w2Indices, w1Indices, a1Ind
 }
 
 // MorphologicalComputationWA2 = I(W;{W,A}) - I(W';A)
-func MorphologicalComputationWA(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationWA(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indices []int, k int, eta bool) []float64 {
 	r1 := FrenzelPompe(w2w1a1, w2Indices, w1Indices, a1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger2(w2w1a1, w2Indices, a1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -55,7 +55,7 @@ func MorphologicalComputationWA(w2w1a1 [][]float64, w2Indices, w1Indices, a1Indi
 }
 
 // MorphologicalComputationWS1 = I(W;{W,S}) - I(W';S)
-func MorphologicalComputationWS1(w2w1s1 [][]float64, w2Indices, w1Indices, s1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationWS1(w2w1s1 [][]float64, w2Indices, w1Indices, s1Indices []int, k int, eta bool) []float64 {
 	r1 := FrenzelPompe(w2w1s1, w2Indices, w1Indices, s1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger1(w2w1s1, w2Indices, s1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -66,7 +66,7 @@ func MorphologicalComputationWS1(w2w1s1 [][]float64, w2Indices, w1Indices, s1Ind
 }
 
 // MorphologicalComputationWS2 = I(W;{W,S}) - I(W';S)
-func MorphologicalComputationWS2(w2w1s1 [][]float64, w2Indices, w1Indices, s1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationWS2(w2w1s1 [][]float64, w2Indices, w1Indices, s1Indices []int, k int, eta bool) []float64 {
 	r1 := FrenzelPompe(w2w1s1, w2Indices, w1Indices, s1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger2(w2w1s1, w2Indices, s1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -77,7 +77,7 @@ func MorphologicalComputationWS2(w2w1s1 [][]float64, w2Indices, w1Indices, s1Ind
 }
 
 // MorphologicalComputationMI1 [...]
-func MorphologicalComputationMI1(w2w1s1a1 [][]float64, w2Indices, w1Indices, s1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationMI1(w2w1s1a1 [][]float64, w2Indices, w1Indices, s1Indices, a1Indices []int, k int, eta bool) []float64 {
 	r1 := KraskovStoegbauerGrassberger1(w2w1s1a1, w2Indices, w1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger1(w2w1s1a1, s1Indices, a1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
@@ -88,7 +88,7 @@ func MorphologicalComputationMI1(w2w1s1a1 [][]float64, w2Indices, w1Indices, s1I
 }
 
 // MorphologicalComputationMI2 [...]
-func MorphologicalComputationMI2(w2w1s1a1 [][]float64, w2Indices, w1Indices, s1Indices, a1Indices []int64, k int64, eta bool) []float64 {
+func MorphologicalComputationMI2(w2w1s1a1 [][]float64, w2Indices, w1Indices, s1Indices, a1Indices []int, k int, eta bool) []float64 {
 	r1 := KraskovStoegbauerGrassberger2(w2w1s1a1, w2Indices, w1Indices, k, eta)
 	r2 := KraskovStoegbauerGrassberger2(w2w1s1a1, a1Indices, s1Indices, k, eta)
 	r := make([]float64, len(r1), len(r2))
